moved Prereq to its own file (to be joined by Course and/or CourseTable objects)
